Crate esp32s3

Peripheral access API for ESP32-S3 microcontrollers (generated using svd2rust v0.31.5 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::i2c0 as i2c1;

pub use self::mcpwm0 as mcpwm1;

pub use self::spi2 as spi3;

pub use self::timg0 as timg1;

pub use self::uart0 as uart1;

pub use self::uart0 as uart2;

pub use self::uhci0 as uhci1;

Modules

aes

AES (Advanced Encryption Standard) Accelerator

apb_ctrl

APB (Advanced Peripheral Bus) Controller

apb_saradc

SAR (Successive Approximation Register) Analog-to-Digital Converter

assist_debug

Debug Assist

bb

BB Peripheral

dma

DMA (Direct Memory Access) Controller

ds

Digital Signature

efuse

eFuse Controller

extmem

External Memory

generic

Common register and bit access and modify traits

gpio

General Purpose Input/Output

gpio_sd

Sigma-Delta Modulation

hmac

HMAC (Hash-based Message Authentication Code) Accelerator

i2c0

I2C (Inter-Integrated Circuit) Controller 0

i2s0

I2S (Inter-IC Sound) Controller 0

i2s1

I2S (Inter-IC Sound) Controller 1

interrupt_core0

Interrupt Controller (Core 0)

interrupt_core1

Interrupt Controller (Core 1)

io_mux

Input/Output Multiplexer

lcd_cam

Camera/LCD Controller

ledc

LED Control PWM (Pulse Width Modulation)

mcpwm0

Motor Control Pulse-Width Modulation 0

pcnt

Pulse Count Controller

peri_backup

PERI_BACKUP Peripheral

rmt

Remote Control

rng

Hardware Random Number Generator

rsa

RSA (Rivest Shamir Adleman) Accelerator

rtc_cntl

Real-Time Clock Control

rtc_i2c

Low-power I2C (Inter-Integrated Circuit) Controller

rtc_io

Low-power Input/Output

sdhost

SD/MMC Host Controller

sens

SENS Peripheral

sensitive

SENSITIVE Peripheral

sha

SHA (Secure Hash Algorithm) Accelerator

spi0

SPI (Serial Peripheral Interface) Controller 0

spi1

SPI (Serial Peripheral Interface) Controller 1

spi2

SPI (Serial Peripheral Interface) Controller 2

system

System Configuration Registers

systimer

System Timer

timg0

Timer Group 0

twai0

Two-Wire Automotive Interface

uart0

UART (Universal Asynchronous Receiver-Transmitter) Controller 0

uhci0

Universal Host Controller Interface 0

usb0

USB OTG (On-The-Go)

usb_device

Full-speed USB Serial/JTAG Controller

usb_wrap

USB_WRAP Peripheral

wcl

WCL Peripheral

xts_aes

XTS-AES-128 Flash Encryption

Structs

AES

AES (Advanced Encryption Standard) Accelerator

APB_CTRL

APB (Advanced Peripheral Bus) Controller

APB_SARADC

SAR (Successive Approximation Register) Analog-to-Digital Converter

ASSIST_DEBUG

Debug Assist

BB

BB Peripheral

DMA

DMA (Direct Memory Access) Controller

DS

Digital Signature

EFUSE

eFuse Controller

EXTMEM

External Memory

GPIO

General Purpose Input/Output

GPIO_SD

Sigma-Delta Modulation

HMAC

HMAC (Hash-based Message Authentication Code) Accelerator

I2C0

I2C (Inter-Integrated Circuit) Controller 0

I2C1

I2C (Inter-Integrated Circuit) Controller 1

I2S0

I2S (Inter-IC Sound) Controller 0

I2S1

I2S (Inter-IC Sound) Controller 1

INTERRUPT_CORE0

Interrupt Controller (Core 0)

INTERRUPT_CORE1

Interrupt Controller (Core 1)

IO_MUX

Input/Output Multiplexer

LCD_CAM

Camera/LCD Controller

LEDC

LED Control PWM (Pulse Width Modulation)

MCPWM0

Motor Control Pulse-Width Modulation 0

MCPWM1

Motor Control Pulse-Width Modulation 1

PCNT

Pulse Count Controller

PERI_BACKUP

PERI_BACKUP Peripheral

Peripherals

All the peripherals.

RMT

Remote Control

RNG

Hardware Random Number Generator

RSA

RSA (Rivest Shamir Adleman) Accelerator

RTC_CNTL

Real-Time Clock Control

RTC_I2C

Low-power I2C (Inter-Integrated Circuit) Controller

RTC_IO

Low-power Input/Output

SDHOST

SD/MMC Host Controller

SENS

SENS Peripheral

SENSITIVE

SENSITIVE Peripheral

SHA

SHA (Secure Hash Algorithm) Accelerator

SPI0

SPI (Serial Peripheral Interface) Controller 0

SPI1

SPI (Serial Peripheral Interface) Controller 1

SPI2

SPI (Serial Peripheral Interface) Controller 2

SPI3

SPI (Serial Peripheral Interface) Controller 3

SYSTEM

System Configuration Registers

SYSTIMER

System Timer

TIMG0

Timer Group 0

TIMG1

Timer Group 1

TWAI0

Two-Wire Automotive Interface

TryFromInterruptError

TryFromInterruptError

UART0

UART (Universal Asynchronous Receiver-Transmitter) Controller 0

UART1

UART (Universal Asynchronous Receiver-Transmitter) Controller 1

UART2

UART (Universal Asynchronous Receiver-Transmitter) Controller 2

UHCI0

Universal Host Controller Interface 0

UHCI1

Universal Host Controller Interface 1

USB0

USB OTG (On-The-Go)

USB_DEVICE

Full-speed USB Serial/JTAG Controller

USB_WRAP

USB_WRAP Peripheral

WCL

WCL Peripheral

XTS_AES

XTS-AES-128 Flash Encryption

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority