Type Alias esp32s3::dma::infifo_status_ch::R

pub type R = R<INFIFO_STATUS_CH_SPEC>;

Register INFIFO_STATUS_CH%s reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn infifo_full_l1(&self) -> INFIFO_FULL_L1_R

Bit 0 - L1 Rx FIFO full signal for Rx channel 0.

pub fn infifo_empty_l1(&self) -> INFIFO_EMPTY_L1_R

Bit 1 - L1 Rx FIFO empty signal for Rx channel 0.

pub fn infifo_full_l2(&self) -> INFIFO_FULL_L2_R

Bit 2 - L2 Rx FIFO full signal for Rx channel 0.

pub fn infifo_empty_l2(&self) -> INFIFO_EMPTY_L2_R

Bit 3 - L2 Rx FIFO empty signal for Rx channel 0.

pub fn infifo_full_l3(&self) -> INFIFO_FULL_L3_R

Bit 4 - L3 Rx FIFO full signal for Rx channel 0.

pub fn infifo_empty_l3(&self) -> INFIFO_EMPTY_L3_R

Bit 5 - L3 Rx FIFO empty signal for Rx channel 0.

pub fn infifo_cnt_l1(&self) -> INFIFO_CNT_L1_R

Bits 6:11 - The register stores the byte number of the data in L1 Rx FIFO for Rx channel 0.

pub fn infifo_cnt_l2(&self) -> INFIFO_CNT_L2_R

Bits 12:18 - The register stores the byte number of the data in L2 Rx FIFO for Rx channel 0.

pub fn infifo_cnt_l3(&self) -> INFIFO_CNT_L3_R

Bits 19:23 - The register stores the byte number of the data in L3 Rx FIFO for Rx channel 0.

pub fn in_remain_under_1b_l3(&self) -> IN_REMAIN_UNDER_1B_L3_R

Bit 24 - reserved

pub fn in_remain_under_2b_l3(&self) -> IN_REMAIN_UNDER_2B_L3_R

Bit 25 - reserved

pub fn in_remain_under_3b_l3(&self) -> IN_REMAIN_UNDER_3B_L3_R

Bit 26 - reserved

pub fn in_remain_under_4b_l3(&self) -> IN_REMAIN_UNDER_4B_L3_R

Bit 27 - reserved

pub fn in_buf_hungry(&self) -> IN_BUF_HUNGRY_R

Bit 28 - reserved