Type Alias esp32s3::dma::outfifo_status_ch::R

pub type R = R<OUTFIFO_STATUS_CH_SPEC>;

Register OUTFIFO_STATUS_CH%s reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn outfifo_full_l1(&self) -> OUTFIFO_FULL_L1_R

Bit 0 - L1 Tx FIFO full signal for Tx channel 0.

pub fn outfifo_empty_l1(&self) -> OUTFIFO_EMPTY_L1_R

Bit 1 - L1 Tx FIFO empty signal for Tx channel 0.

pub fn outfifo_full_l2(&self) -> OUTFIFO_FULL_L2_R

Bit 2 - L2 Tx FIFO full signal for Tx channel 0.

pub fn outfifo_empty_l2(&self) -> OUTFIFO_EMPTY_L2_R

Bit 3 - L2 Tx FIFO empty signal for Tx channel 0.

pub fn outfifo_full_l3(&self) -> OUTFIFO_FULL_L3_R

Bit 4 - L3 Tx FIFO full signal for Tx channel 0.

pub fn outfifo_empty_l3(&self) -> OUTFIFO_EMPTY_L3_R

Bit 5 - L3 Tx FIFO empty signal for Tx channel 0.

pub fn outfifo_cnt_l1(&self) -> OUTFIFO_CNT_L1_R

Bits 6:10 - The register stores the byte number of the data in L1 Tx FIFO for Tx channel 0.

pub fn outfifo_cnt_l2(&self) -> OUTFIFO_CNT_L2_R

Bits 11:17 - The register stores the byte number of the data in L2 Tx FIFO for Tx channel 0.

pub fn outfifo_cnt_l3(&self) -> OUTFIFO_CNT_L3_R

Bits 18:22 - The register stores the byte number of the data in L3 Tx FIFO for Tx channel 0.

pub fn out_remain_under_1b_l3(&self) -> OUT_REMAIN_UNDER_1B_L3_R

Bit 23 - reserved

pub fn out_remain_under_2b_l3(&self) -> OUT_REMAIN_UNDER_2B_L3_R

Bit 24 - reserved

pub fn out_remain_under_3b_l3(&self) -> OUT_REMAIN_UNDER_3B_L3_R

Bit 25 - reserved

pub fn out_remain_under_4b_l3(&self) -> OUT_REMAIN_UNDER_4B_L3_R

Bit 26 - reserved