Type Alias esp32s3::efuse::rd_repeat_data2::R

pub type R = R<RD_REPEAT_DATA2_SPEC>;

Register RD_REPEAT_DATA2 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn key_purpose_2(&self) -> KEY_PURPOSE_2_R

Bits 0:3 - Purpose of Key2.

pub fn key_purpose_3(&self) -> KEY_PURPOSE_3_R

Bits 4:7 - Purpose of Key3.

pub fn key_purpose_4(&self) -> KEY_PURPOSE_4_R

Bits 8:11 - Purpose of Key4.

pub fn key_purpose_5(&self) -> KEY_PURPOSE_5_R

Bits 12:15 - Purpose of Key5.

pub fn rpt4_reserved0(&self) -> RPT4_RESERVED0_R

Bits 16:19 - Reserved (used for four backups method).

pub fn secure_boot_en(&self) -> SECURE_BOOT_EN_R

Bit 20 - Set this bit to enable secure boot.

pub fn secure_boot_aggressive_revoke(&self) -> SECURE_BOOT_AGGRESSIVE_REVOKE_R

Bit 21 - Set this bit to enable revoking aggressive secure boot.

pub fn dis_usb_jtag(&self) -> DIS_USB_JTAG_R

Bit 22 - Set this bit to disable function of usb switch to jtag in module of usb device.

pub fn dis_usb_device(&self) -> DIS_USB_DEVICE_R

Bit 23 - Set this bit to disable usb device.

pub fn strap_jtag_sel(&self) -> STRAP_JTAG_SEL_R

Bit 24 - Set this bit to enable selection between usb_to_jtag and pad_to_jtag through strapping gpio10 when both reg_dis_usb_jtag and reg_dis_pad_jtag are equal to 0.

pub fn usb_phy_sel(&self) -> USB_PHY_SEL_R

Bit 25 - This bit is used to switch internal PHY and external PHY for USB OTG and USB Device. 0: internal PHY is assigned to USB Device while external PHY is assigned to USB OTG. 1: internal PHY is assigned to USB OTG while external PHY is assigned to USB Device.

pub fn power_glitch_dsense(&self) -> POWER_GLITCH_DSENSE_R

Bits 26:27 - Sample delay configuration of power glitch.

pub fn flash_tpuw(&self) -> FLASH_TPUW_R

Bits 28:31 - Configures flash waiting time after power-up, in unit of ms. If the value is less than 15, the waiting time is the configurable value. Otherwise, the waiting time is twice the configurable value.