Type Alias esp32s3::lcd_cam::cam_ctrl1::R

pub type R = R<CAM_CTRL1_SPEC>;

Register CAM_CTRL1 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn cam_rec_data_bytelen(&self) -> CAM_REC_DATA_BYTELEN_R

Bits 0:15 - Configure camera received data byte length. When the length of received data reaches this value + 1, GDMA in_suc_eof_int is triggered.

pub fn cam_line_int_num(&self) -> CAM_LINE_INT_NUM_R

Bits 16:21 - Configure line number. When the number of received lines reaches this value + 1, LCD_CAM_CAM_HS_INT is triggered.

pub fn cam_clk_inv(&self) -> CAM_CLK_INV_R

Bit 22 - 1: Invert the input signal CAM_PCLK. 0: Do not invert.

pub fn cam_vsync_filter_en(&self) -> CAM_VSYNC_FILTER_EN_R

Bit 23 - 1: Enable CAM_VSYNC filter function. 0: Bypass.

pub fn cam_2byte_en(&self) -> CAM_2BYTE_EN_R

Bit 24 - 1: The width of input data is 16 bits. 0: The width of input data is 8 bits.

pub fn cam_de_inv(&self) -> CAM_DE_INV_R

Bit 25 - CAM_DE invert enable signal, valid in high level.

pub fn cam_hsync_inv(&self) -> CAM_HSYNC_INV_R

Bit 26 - CAM_HSYNC invert enable signal, valid in high level.

pub fn cam_vsync_inv(&self) -> CAM_VSYNC_INV_R

Bit 27 - CAM_VSYNC invert enable signal, valid in high level.

pub fn cam_vh_de_mode_en(&self) -> CAM_VH_DE_MODE_EN_R

Bit 28 - 1: Input control signals are CAM_DE and CAM_HSYNC. CAM_VSYNC is 1. 0: Input control signals are CAM_DE and CAM_VSYNC. CAM_HSYNC and CAM_DE are all 1 at the the same time.

pub fn cam_start(&self) -> CAM_START_R

Bit 29 - Camera module start signal.