Module esp32s3::lcd_cam::lcd_misc

LCD MISC configuration register

Structs

LCD_MISC_SPEC

LCD MISC configuration register

Type Aliases

LCD_AFIFO_RESET_W

Field LCD_AFIFO_RESET writer - Async Tx FIFO reset signal.

LCD_AFIFO_THRESHOLD_NUM_R

Field LCD_AFIFO_THRESHOLD_NUM reader - Set the threshold for Async Tx FIFO full event.

LCD_AFIFO_THRESHOLD_NUM_W

Field LCD_AFIFO_THRESHOLD_NUM writer - Set the threshold for Async Tx FIFO full event.

LCD_BK_EN_R

Field LCD_BK_EN reader - 1: Enable blank region when LCD sends data out. 0: No blank region.

LCD_BK_EN_W

Field LCD_BK_EN writer - 1: Enable blank region when LCD sends data out. 0: No blank region.

LCD_CD_CMD_SET_R

Field LCD_CD_CMD_SET reader - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in CMD phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

LCD_CD_CMD_SET_W

Field LCD_CD_CMD_SET writer - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in CMD phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

LCD_CD_DATA_SET_R

Field LCD_CD_DATA_SET reader - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in DOUT phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

LCD_CD_DATA_SET_W

Field LCD_CD_DATA_SET writer - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in DOUT phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

LCD_CD_DUMMY_SET_R

Field LCD_CD_DUMMY_SET reader - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in DUMMY phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

LCD_CD_DUMMY_SET_W

Field LCD_CD_DUMMY_SET writer - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in DUMMY phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

LCD_CD_IDLE_EDGE_R

Field LCD_CD_IDLE_EDGE reader - The default value of LCD_CD.

LCD_CD_IDLE_EDGE_W

Field LCD_CD_IDLE_EDGE writer - The default value of LCD_CD.

LCD_NEXT_FRAME_EN_R

Field LCD_NEXT_FRAME_EN reader - 1: Send the next frame data when the current frame is sent out. 0: LCD stops when the current frame is sent out.

LCD_NEXT_FRAME_EN_W

Field LCD_NEXT_FRAME_EN writer - 1: Send the next frame data when the current frame is sent out. 0: LCD stops when the current frame is sent out.

LCD_VBK_CYCLELEN_R

Field LCD_VBK_CYCLELEN reader - Configure the hold time cycles in LCD non-RGB mode. Hold cycles expected = this value + 1. %Configure the cycles for vertical back blank region in LCD RGB mode, the cycles = this value + 1. Or configure the hold time cycles in LCD non-RGB mode, the cycles = this value + 1.

LCD_VBK_CYCLELEN_W

Field LCD_VBK_CYCLELEN writer - Configure the hold time cycles in LCD non-RGB mode. Hold cycles expected = this value + 1. %Configure the cycles for vertical back blank region in LCD RGB mode, the cycles = this value + 1. Or configure the hold time cycles in LCD non-RGB mode, the cycles = this value + 1.

LCD_VFK_CYCLELEN_R

Field LCD_VFK_CYCLELEN reader - Configure the setup cycles in LCD non-RGB mode. Setup cycles expected = this value + 1.

LCD_VFK_CYCLELEN_W

Field LCD_VFK_CYCLELEN writer - Configure the setup cycles in LCD non-RGB mode. Setup cycles expected = this value + 1.

R

Register LCD_MISC reader

W

Register LCD_MISC writer