Type Alias esp32s3::lcd_cam::lcd_misc::R

pub type R = R<LCD_MISC_SPEC>;

Register LCD_MISC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn lcd_afifo_threshold_num(&self) -> LCD_AFIFO_THRESHOLD_NUM_R

Bits 1:5 - Set the threshold for Async Tx FIFO full event.

pub fn lcd_vfk_cyclelen(&self) -> LCD_VFK_CYCLELEN_R

Bits 6:11 - Configure the setup cycles in LCD non-RGB mode. Setup cycles expected = this value + 1.

pub fn lcd_vbk_cyclelen(&self) -> LCD_VBK_CYCLELEN_R

Bits 12:24 - Configure the hold time cycles in LCD non-RGB mode. Hold cycles expected = this value + 1. %Configure the cycles for vertical back blank region in LCD RGB mode, the cycles = this value + 1. Or configure the hold time cycles in LCD non-RGB mode, the cycles = this value + 1.

pub fn lcd_next_frame_en(&self) -> LCD_NEXT_FRAME_EN_R

Bit 25 - 1: Send the next frame data when the current frame is sent out. 0: LCD stops when the current frame is sent out.

pub fn lcd_bk_en(&self) -> LCD_BK_EN_R

Bit 26 - 1: Enable blank region when LCD sends data out. 0: No blank region.

pub fn lcd_cd_data_set(&self) -> LCD_CD_DATA_SET_R

Bit 28 - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in DOUT phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

pub fn lcd_cd_dummy_set(&self) -> LCD_CD_DUMMY_SET_R

Bit 29 - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in DUMMY phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

pub fn lcd_cd_cmd_set(&self) -> LCD_CD_CMD_SET_R

Bit 30 - 1: LCD_CD = !LCD_CAM_LCD_CD_IDLE_EDGE when LCD is in CMD phase. 0: LCD_CD = LCD_CAM_LCD_CD_IDLE_EDGE.

pub fn lcd_cd_idle_edge(&self) -> LCD_CD_IDLE_EDGE_R

Bit 31 - The default value of LCD_CD.