Module esp32s3::lcd_cam::lcd_user

LCD user configuration register

Structs

LCD_USER_SPEC

LCD user configuration register

Type Aliases

LCD_2BYTE_EN_R

Field LCD_2BYTE_EN reader - 1: The width of output LCD data is 16 bits. 0: The width of output LCD data is 8 bits.

LCD_2BYTE_EN_W

Field LCD_2BYTE_EN writer - 1: The width of output LCD data is 16 bits. 0: The width of output LCD data is 8 bits.

LCD_8BITS_ORDER_R

Field LCD_8BITS_ORDER reader - 1: Swap every two data bytes, valid in 8-bit mode. 0: Do not swap.

LCD_8BITS_ORDER_W

Field LCD_8BITS_ORDER writer - 1: Swap every two data bytes, valid in 8-bit mode. 0: Do not swap.

LCD_ALWAYS_OUT_EN_R

Field LCD_ALWAYS_OUT_EN reader - LCD continues outputting data when LCD is in DOUT phase, till LCD_CAM_LCD_START is cleared or LCD_CAM_LCD_RESET is set.

LCD_ALWAYS_OUT_EN_W

Field LCD_ALWAYS_OUT_EN writer - LCD continues outputting data when LCD is in DOUT phase, till LCD_CAM_LCD_START is cleared or LCD_CAM_LCD_RESET is set.

LCD_BIT_ORDER_R

Field LCD_BIT_ORDER reader - 1: Change data bit order. Change LCD_DATA_out[7:0] to LCD_DATA_out[0:7] in 8-bit mode, and bits[15:0] to bits[0:15] in 16-bit mode. 0: Do not change.

LCD_BIT_ORDER_W

Field LCD_BIT_ORDER writer - 1: Change data bit order. Change LCD_DATA_out[7:0] to LCD_DATA_out[0:7] in 8-bit mode, and bits[15:0] to bits[0:15] in 16-bit mode. 0: Do not change.

LCD_BYTE_ORDER_R

Field LCD_BYTE_ORDER reader - 1: Invert data byte order, only valid in 16-bit mode. 0: Do not invert.

LCD_BYTE_ORDER_W

Field LCD_BYTE_ORDER writer - 1: Invert data byte order, only valid in 16-bit mode. 0: Do not invert.

LCD_CMD_2_CYCLE_EN_R

Field LCD_CMD_2_CYCLE_EN reader - The cycle length of command phase. 1: 2 cycles. 0: 1 cycle.

LCD_CMD_2_CYCLE_EN_W

Field LCD_CMD_2_CYCLE_EN writer - The cycle length of command phase. 1: 2 cycles. 0: 1 cycle.

LCD_CMD_R

Field LCD_CMD reader - 1: Be able to send command in LCD sequence when LCD starts. 0: Disable.

LCD_CMD_W

Field LCD_CMD writer - 1: Be able to send command in LCD sequence when LCD starts. 0: Disable.

LCD_DOUT_CYCLELEN_R

Field LCD_DOUT_CYCLELEN reader - Configure the cycles for DOUT phase of LCD module. The cycles = this value + 1.

LCD_DOUT_CYCLELEN_W

Field LCD_DOUT_CYCLELEN writer - Configure the cycles for DOUT phase of LCD module. The cycles = this value + 1.

LCD_DOUT_R

Field LCD_DOUT reader - 1: Be able to send data out in LCD sequence when LCD starts. 0: Disable.

LCD_DOUT_W

Field LCD_DOUT writer - 1: Be able to send data out in LCD sequence when LCD starts. 0: Disable.

LCD_DUMMY_CYCLELEN_R

Field LCD_DUMMY_CYCLELEN reader - Configure DUMMY cycles. DUMMY cycles = this value + 1.

LCD_DUMMY_CYCLELEN_W

Field LCD_DUMMY_CYCLELEN writer - Configure DUMMY cycles. DUMMY cycles = this value + 1.

LCD_DUMMY_R

Field LCD_DUMMY reader - 1: Enable DUMMY phase in LCD sequence when LCD starts. 0: Disable.

LCD_DUMMY_W

Field LCD_DUMMY writer - 1: Enable DUMMY phase in LCD sequence when LCD starts. 0: Disable.

LCD_RESET_W

Field LCD_RESET writer - Reset LCD module.

LCD_START_R

Field LCD_START reader - LCD starts sending data enable signal, valid in high level.

LCD_START_W

Field LCD_START writer - LCD starts sending data enable signal, valid in high level.

LCD_UPDATE_R

Field LCD_UPDATE reader - 1: Update LCD registers. This bit is cleared by hardware. 0: Do not care.

LCD_UPDATE_W

Field LCD_UPDATE writer - 1: Update LCD registers. This bit is cleared by hardware. 0: Do not care.

R

Register LCD_USER reader

W

Register LCD_USER writer