Module esp32s3::mcpwm0

Motor Control Pulse-Width Modulation 0

Modules

cap_ch0

Value of last capture on channel 0

cap_ch0_cfg

Capture channel 0 configuration and enable

cap_ch1

Value of last capture on channel 1

cap_ch1_cfg

Capture channel 1 configuration and enable

cap_ch2

Value of last capture on channel 2

cap_ch2_cfg

Capture channel 2 configuration and enable

cap_status

Edge of last capture trigger

cap_timer_cfg

Configure capture timer

cap_timer_phase

Phase for capture timer sync

chopper0_cfg

Carrier enable and configuratoin

chopper1_cfg

Carrier enable and configuratoin

chopper2_cfg

Carrier enable and configuratoin

clk

MCPWM APB configuration register

clk_cfg

PWM clock prescaler register.

cmpr0_cfg

Transfer status and update method for time stamp registers A and B

cmpr0_value0

Shadow register for register A.

cmpr0_value1

Shadow register for register B.

cmpr1_cfg

Transfer status and update method for time stamp registers A and B

cmpr1_value0

Shadow register for register A.

cmpr1_value1

Shadow register for register B.

cmpr2_cfg

Transfer status and update method for time stamp registers A and B

cmpr2_value0

Shadow register for register A.

cmpr2_value1

Shadow register for register B.

db0_cfg

dead time type selection and configuration

db0_fed_cfg

Shadow register for falling edge delay (FED).

db0_red_cfg

Shadow register for rising edge delay (RED).

db1_cfg

dead time type selection and configuration

db1_fed_cfg

Shadow register for falling edge delay (FED).

db1_red_cfg

Shadow register for rising edge delay (RED).

db2_cfg

dead time type selection and configuration

db2_fed_cfg

Shadow register for falling edge delay (FED).

db2_red_cfg

Shadow register for rising edge delay (RED).

fault_detect

Fault detection configuration and status

gen0_a

Actions triggered by events on PWM0A

gen0_b

Actions triggered by events on PWM0B

gen0_cfg0

Fault event T0 and T1 handling

gen0_force

Permissives to force PWM0A and PWM0B outputs by software

gen1_a

Actions triggered by events on PWM1A

gen1_b

Actions triggered by events on PWM1B

gen1_cfg0

Fault event T0 and T1 handling

gen1_force

Permissives to force PWM1A and PWM1B outputs by software

gen2_a

Actions triggered by events on PWM2A

gen2_b

Actions triggered by events on PWM2B

gen2_cfg0

Fault event T0 and T1 handling

gen2_force

Permissives to force PWM2A and PWM2B outputs by software

int_clr

Interrupt clear bits

int_ena

Interrupt enable bits

int_raw

Raw interrupt status

int_st

Masked interrupt status

operator_timersel

Select specific timer for PWM operators.

timer0_cfg0

PWM timer0 period and update method configuration register.

timer0_cfg1

PWM timer0 working mode and start/stop control configuration register.

timer0_status

PWM timer0 status register.

timer0_sync

PWM timer0 sync function configuration register.

timer1_cfg0

PWM timer1 period and update method configuration register.

timer1_cfg1

PWM timer1 working mode and start/stop control configuration register.

timer1_status

PWM timer1 status register.

timer1_sync

PWM timer1 sync function configuration register.

timer2_cfg0

PWM timer2 period and update method configuration register.

timer2_cfg1

PWM timer2 working mode and start/stop control configuration register.

timer2_status

PWM timer2 status register.

timer2_sync

PWM timer2 sync function configuration register.

timer_synci_cfg

Synchronization input selection for three PWM timers.

tz0_cfg0

Actions on PWM0A and PWM0B trip events

tz0_cfg1

Software triggers for fault handler actions

tz0_status

Status of fault events.

tz1_cfg0

Actions on PWM1A and PWM1B trip events

tz1_cfg1

Software triggers for fault handler actions

tz1_status

Status of fault events.

tz2_cfg0

Actions on PWM2A and PWM2B trip events

tz2_cfg1

Software triggers for fault handler actions

tz2_status

Status of fault events.

update_cfg

Enable update.

version

Version register.

Structs

RegisterBlock

Register block

Type Aliases

CAP_CH0

CAP_CH0 (r) register accessor: Value of last capture on channel 0

CAP_CH0_CFG

CAP_CH0_CFG (rw) register accessor: Capture channel 0 configuration and enable

CAP_CH1

CAP_CH1 (r) register accessor: Value of last capture on channel 1

CAP_CH1_CFG

CAP_CH1_CFG (rw) register accessor: Capture channel 1 configuration and enable

CAP_CH2

CAP_CH2 (r) register accessor: Value of last capture on channel 2

CAP_CH2_CFG

CAP_CH2_CFG (rw) register accessor: Capture channel 2 configuration and enable

CAP_STATUS

CAP_STATUS (r) register accessor: Edge of last capture trigger

CAP_TIMER_CFG

CAP_TIMER_CFG (rw) register accessor: Configure capture timer

CAP_TIMER_PHASE

CAP_TIMER_PHASE (rw) register accessor: Phase for capture timer sync

CHOPPER0_CFG

CHOPPER0_CFG (rw) register accessor: Carrier enable and configuratoin

CHOPPER1_CFG

CHOPPER1_CFG (rw) register accessor: Carrier enable and configuratoin

CHOPPER2_CFG

CHOPPER2_CFG (rw) register accessor: Carrier enable and configuratoin

CLK

CLK (rw) register accessor: MCPWM APB configuration register

CLK_CFG

CLK_CFG (rw) register accessor: PWM clock prescaler register.

CMPR0_CFG

CMPR0_CFG (rw) register accessor: Transfer status and update method for time stamp registers A and B

CMPR0_VALUE0

CMPR0_VALUE0 (rw) register accessor: Shadow register for register A.

CMPR0_VALUE1

CMPR0_VALUE1 (rw) register accessor: Shadow register for register B.

CMPR1_CFG

CMPR1_CFG (rw) register accessor: Transfer status and update method for time stamp registers A and B

CMPR1_VALUE0

CMPR1_VALUE0 (rw) register accessor: Shadow register for register A.

CMPR1_VALUE1

CMPR1_VALUE1 (rw) register accessor: Shadow register for register B.

CMPR2_CFG

CMPR2_CFG (rw) register accessor: Transfer status and update method for time stamp registers A and B

CMPR2_VALUE0

CMPR2_VALUE0 (rw) register accessor: Shadow register for register A.

CMPR2_VALUE1

CMPR2_VALUE1 (rw) register accessor: Shadow register for register B.

DB0_CFG

DB0_CFG (rw) register accessor: dead time type selection and configuration

DB0_FED_CFG

DB0_FED_CFG (rw) register accessor: Shadow register for falling edge delay (FED).

DB0_RED_CFG

DB0_RED_CFG (rw) register accessor: Shadow register for rising edge delay (RED).

DB1_CFG

DB1_CFG (rw) register accessor: dead time type selection and configuration

DB1_FED_CFG

DB1_FED_CFG (rw) register accessor: Shadow register for falling edge delay (FED).

DB1_RED_CFG

DB1_RED_CFG (rw) register accessor: Shadow register for rising edge delay (RED).

DB2_CFG

DB2_CFG (rw) register accessor: dead time type selection and configuration

DB2_FED_CFG

DB2_FED_CFG (rw) register accessor: Shadow register for falling edge delay (FED).

DB2_RED_CFG

DB2_RED_CFG (rw) register accessor: Shadow register for rising edge delay (RED).

FAULT_DETECT

FAULT_DETECT (rw) register accessor: Fault detection configuration and status

GEN0_A

GEN0_A (rw) register accessor: Actions triggered by events on PWM0A

GEN0_B

GEN0_B (rw) register accessor: Actions triggered by events on PWM0B

GEN0_CFG0

GEN0_CFG0 (rw) register accessor: Fault event T0 and T1 handling

GEN0_FORCE

GEN0_FORCE (rw) register accessor: Permissives to force PWM0A and PWM0B outputs by software

GEN1_A

GEN1_A (rw) register accessor: Actions triggered by events on PWM1A

GEN1_B

GEN1_B (rw) register accessor: Actions triggered by events on PWM1B

GEN1_CFG0

GEN1_CFG0 (rw) register accessor: Fault event T0 and T1 handling

GEN1_FORCE

GEN1_FORCE (rw) register accessor: Permissives to force PWM1A and PWM1B outputs by software

GEN2_A

GEN2_A (rw) register accessor: Actions triggered by events on PWM2A

GEN2_B

GEN2_B (rw) register accessor: Actions triggered by events on PWM2B

GEN2_CFG0

GEN2_CFG0 (rw) register accessor: Fault event T0 and T1 handling

GEN2_FORCE

GEN2_FORCE (rw) register accessor: Permissives to force PWM2A and PWM2B outputs by software

INT_CLR

INT_CLR (w) register accessor: Interrupt clear bits

INT_ENA

INT_ENA (rw) register accessor: Interrupt enable bits

INT_RAW

INT_RAW (rw) register accessor: Raw interrupt status

INT_ST

INT_ST (r) register accessor: Masked interrupt status

OPERATOR_TIMERSEL

OPERATOR_TIMERSEL (rw) register accessor: Select specific timer for PWM operators.

TIMER0_CFG0

TIMER0_CFG0 (rw) register accessor: PWM timer0 period and update method configuration register.

TIMER0_CFG1

TIMER0_CFG1 (rw) register accessor: PWM timer0 working mode and start/stop control configuration register.

TIMER0_STATUS

TIMER0_STATUS (r) register accessor: PWM timer0 status register.

TIMER0_SYNC

TIMER0_SYNC (rw) register accessor: PWM timer0 sync function configuration register.

TIMER1_CFG0

TIMER1_CFG0 (rw) register accessor: PWM timer1 period and update method configuration register.

TIMER1_CFG1

TIMER1_CFG1 (rw) register accessor: PWM timer1 working mode and start/stop control configuration register.

TIMER1_STATUS

TIMER1_STATUS (r) register accessor: PWM timer1 status register.

TIMER1_SYNC

TIMER1_SYNC (rw) register accessor: PWM timer1 sync function configuration register.

TIMER2_CFG0

TIMER2_CFG0 (rw) register accessor: PWM timer2 period and update method configuration register.

TIMER2_CFG1

TIMER2_CFG1 (rw) register accessor: PWM timer2 working mode and start/stop control configuration register.

TIMER2_STATUS

TIMER2_STATUS (r) register accessor: PWM timer2 status register.

TIMER2_SYNC

TIMER2_SYNC (rw) register accessor: PWM timer2 sync function configuration register.

TIMER_SYNCI_CFG

TIMER_SYNCI_CFG (rw) register accessor: Synchronization input selection for three PWM timers.

TZ0_CFG0

TZ0_CFG0 (rw) register accessor: Actions on PWM0A and PWM0B trip events

TZ0_CFG1

TZ0_CFG1 (rw) register accessor: Software triggers for fault handler actions

TZ0_STATUS

TZ0_STATUS (r) register accessor: Status of fault events.

TZ1_CFG0

TZ1_CFG0 (rw) register accessor: Actions on PWM1A and PWM1B trip events

TZ1_CFG1

TZ1_CFG1 (rw) register accessor: Software triggers for fault handler actions

TZ1_STATUS

TZ1_STATUS (r) register accessor: Status of fault events.

TZ2_CFG0

TZ2_CFG0 (rw) register accessor: Actions on PWM2A and PWM2B trip events

TZ2_CFG1

TZ2_CFG1 (rw) register accessor: Software triggers for fault handler actions

TZ2_STATUS

TZ2_STATUS (r) register accessor: Status of fault events.

UPDATE_CFG

UPDATE_CFG (rw) register accessor: Enable update.

VERSION

VERSION (rw) register accessor: Version register.