Module esp32s3::mcpwm0::chopper2_cfg

Carrier enable and configuratoin

Structs

CHOPPER2_CFG_SPEC

Carrier enable and configuratoin

Type Aliases

CHOPPER2_DUTY_R

Field CHOPPER2_DUTY reader - carrier duty selection. Duty = PWM_CARRIER0_DUTY / 8

CHOPPER2_DUTY_W

Field CHOPPER2_DUTY writer - carrier duty selection. Duty = PWM_CARRIER0_DUTY / 8

CHOPPER2_EN_R

Field CHOPPER2_EN reader - When set, carrier0 function is enabled. When cleared, carrier0 is bypassed

CHOPPER2_EN_W

Field CHOPPER2_EN writer - When set, carrier0 function is enabled. When cleared, carrier0 is bypassed

CHOPPER2_IN_INVERT_R

Field CHOPPER2_IN_INVERT reader - when set, invert the input of PWM2A and PWM2B for this submodule

CHOPPER2_IN_INVERT_W

Field CHOPPER2_IN_INVERT writer - when set, invert the input of PWM2A and PWM2B for this submodule

CHOPPER2_OSHTWTH_R

Field CHOPPER2_OSHTWTH reader - width of the fist pulse in number of periods of the carrier

CHOPPER2_OSHTWTH_W

Field CHOPPER2_OSHTWTH writer - width of the fist pulse in number of periods of the carrier

CHOPPER2_OUT_INVERT_R

Field CHOPPER2_OUT_INVERT reader - when set, invert the output of PWM2A and PWM2B for this submodule

CHOPPER2_OUT_INVERT_W

Field CHOPPER2_OUT_INVERT writer - when set, invert the output of PWM2A and PWM2B for this submodule

CHOPPER2_PRESCALE_R

Field CHOPPER2_PRESCALE reader - PWM carrier0 clock (PC_clk) prescale value. Period of PC_clk = period of PWM_clk * (PWM_CARRIER0_PRESCALE + 1)

CHOPPER2_PRESCALE_W

Field CHOPPER2_PRESCALE writer - PWM carrier0 clock (PC_clk) prescale value. Period of PC_clk = period of PWM_clk * (PWM_CARRIER0_PRESCALE + 1)

R

Register CHOPPER2_CFG reader

W

Register CHOPPER2_CFG writer