Type Alias esp32s3::mcpwm0::tz1_cfg0::W

pub type W = W<TZ1_CFG0_SPEC>;

Register TZ1_CFG0 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn tz1_sw_cbc(&mut self) -> TZ1_SW_CBC_W<'_, TZ1_CFG0_SPEC>

Bit 0 - Enable register for software force cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_f2_cbc(&mut self) -> TZ1_F2_CBC_W<'_, TZ1_CFG0_SPEC>

Bit 1 - event_f2 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_f1_cbc(&mut self) -> TZ1_F1_CBC_W<'_, TZ1_CFG0_SPEC>

Bit 2 - event_f1 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_f0_cbc(&mut self) -> TZ1_F0_CBC_W<'_, TZ1_CFG0_SPEC>

Bit 3 - event_f0 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_sw_ost(&mut self) -> TZ1_SW_OST_W<'_, TZ1_CFG0_SPEC>

Bit 4 - Enable register for software force one-shot mode action. 0: disable, 1: enable

pub fn tz1_f2_ost(&mut self) -> TZ1_F2_OST_W<'_, TZ1_CFG0_SPEC>

Bit 5 - event_f2 will trigger one-shot mode action. 0: disable, 1: enable

pub fn tz1_f1_ost(&mut self) -> TZ1_F1_OST_W<'_, TZ1_CFG0_SPEC>

Bit 6 - event_f1 will trigger one-shot mode action. 0: disable, 1: enable

pub fn tz1_f0_ost(&mut self) -> TZ1_F0_OST_W<'_, TZ1_CFG0_SPEC>

Bit 7 - event_f0 will trigger one-shot mode action. 0: disable, 1: enable

pub fn tz1_a_cbc_d(&mut self) -> TZ1_A_CBC_D_W<'_, TZ1_CFG0_SPEC>

Bits 8:9 - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_a_cbc_u(&mut self) -> TZ1_A_CBC_U_W<'_, TZ1_CFG0_SPEC>

Bits 10:11 - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_a_ost_d(&mut self) -> TZ1_A_OST_D_W<'_, TZ1_CFG0_SPEC>

Bits 12:13 - One-shot mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_a_ost_u(&mut self) -> TZ1_A_OST_U_W<'_, TZ1_CFG0_SPEC>

Bits 14:15 - One-shot mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_b_cbc_d(&mut self) -> TZ1_B_CBC_D_W<'_, TZ1_CFG0_SPEC>

Bits 16:17 - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_b_cbc_u(&mut self) -> TZ1_B_CBC_U_W<'_, TZ1_CFG0_SPEC>

Bits 18:19 - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_b_ost_d(&mut self) -> TZ1_B_OST_D_W<'_, TZ1_CFG0_SPEC>

Bits 20:21 - One-shot mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub fn tz1_b_ost_u(&mut self) -> TZ1_B_OST_U_W<'_, TZ1_CFG0_SPEC>

Bits 22:23 - One-shot mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force lo, 2: force hi, 3: toggle

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual