Type Alias esp32s3::mcpwm0::update_cfg::R

pub type R = R<UPDATE_CFG_SPEC>;

Register UPDATE_CFG reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn global_up_en(&self) -> GLOBAL_UP_EN_R

Bit 0 - The global enable of update of all active registers in MCPWM module

pub fn global_force_up(&self) -> GLOBAL_FORCE_UP_R

Bit 1 - a toggle (software invert its value) will trigger a forced update of all active registers in MCPWM module

pub fn op0_up_en(&self) -> OP0_UP_EN_R

Bit 2 - When set and PWM_GLOBAL_UP_EN is set, update of active registers in PWM operator 0 are enabled

pub fn op0_force_up(&self) -> OP0_FORCE_UP_R

Bit 3 - a toggle (software invert its value) will trigger a forced update of active registers in PWM operator 0

pub fn op1_up_en(&self) -> OP1_UP_EN_R

Bit 4 - When set and PWM_GLOBAL_UP_EN is set, update of active registers in PWM operator 1 are enabled

pub fn op1_force_up(&self) -> OP1_FORCE_UP_R

Bit 5 - a toggle (software invert its value) will trigger a forced update of active registers in PWM operator 1

pub fn op2_up_en(&self) -> OP2_UP_EN_R

Bit 6 - When set and PWM_GLOBAL_UP_EN is set, update of active registers in PWM operator 2 are enabled

pub fn op2_force_up(&self) -> OP2_FORCE_UP_R

Bit 7 - a toggle (software invert its value) will trigger a forced update of active registers in PWM operator 2