Type Alias esp32s3::pcnt::u_status::R

pub type R = R<U_STATUS_SPEC>;

Register U%s_STATUS reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn zero_mode(&self) -> ZERO_MODE_R

Bits 0:1 - The pulse counter status of PCNT_U%s corresponding to 0. 0: pulse counter decreases from positive to 0. 1: pulse counter increases from negative to 0. 2: pulse counter is negative. 3: pulse counter is positive.

pub fn thres1(&self) -> THRES1_R

Bit 2 - The latched value of thres1 event of PCNT_U%s when threshold event interrupt is valid. 1: the current pulse counter equals to thres1 and thres1 event is valid. 0: others

pub fn thres0(&self) -> THRES0_R

Bit 3 - The latched value of thres0 event of PCNT_U%s when threshold event interrupt is valid. 1: the current pulse counter equals to thres0 and thres0 event is valid. 0: others

pub fn l_lim(&self) -> L_LIM_R

Bit 4 - The latched value of low limit event of PCNT_U%s when threshold event interrupt is valid. 1: the current pulse counter equals to thr_l_lim and low limit event is valid. 0: others

pub fn h_lim(&self) -> H_LIM_R

Bit 5 - The latched value of high limit event of PCNT_U%s when threshold event interrupt is valid. 1: the current pulse counter equals to thr_h_lim and high limit event is valid. 0: others

pub fn zero(&self) -> ZERO_R

Bit 6 - The latched value of zero threshold event of PCNT_U%s when threshold event interrupt is valid. 1: the current pulse counter equals to 0 and zero threshold event is valid. 0: others