Type Alias esp32s3::rmt::int_raw::R

pub type R = R<INT_RAW_SPEC>;

Register INT_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ch_tx_end(&self, n: u8) -> CH_TX_END_R

The interrupt raw bit for CHANNEL(0-3). Triggered when transmission done.

NOTE: n is number of field in register. n == 0 corresponds to CH0_TX_END field

pub fn ch_tx_end_iter(&self) -> impl Iterator<Item = CH_TX_END_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL(0-3). Triggered when transmission done.

pub fn ch0_tx_end(&self) -> CH_TX_END_R

Bit 0 - The interrupt raw bit for CHANNEL0. Triggered when transmission done.

pub fn ch1_tx_end(&self) -> CH_TX_END_R

Bit 1 - The interrupt raw bit for CHANNEL1. Triggered when transmission done.

pub fn ch2_tx_end(&self) -> CH_TX_END_R

Bit 2 - The interrupt raw bit for CHANNEL2. Triggered when transmission done.

pub fn ch3_tx_end(&self) -> CH_TX_END_R

Bit 3 - The interrupt raw bit for CHANNEL3. Triggered when transmission done.

pub fn ch_tx_err(&self, n: u8) -> CH_TX_ERR_R

The interrupt raw bit for CHANNEL(0-3). Triggered when error occurs.

NOTE: n is number of field in register. n == 0 corresponds to CH0_TX_ERR field

pub fn ch_tx_err_iter(&self) -> impl Iterator<Item = CH_TX_ERR_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL(0-3). Triggered when error occurs.

pub fn ch0_tx_err(&self) -> CH_TX_ERR_R

Bit 4 - The interrupt raw bit for CHANNEL0. Triggered when error occurs.

pub fn ch1_tx_err(&self) -> CH_TX_ERR_R

Bit 5 - The interrupt raw bit for CHANNEL1. Triggered when error occurs.

pub fn ch2_tx_err(&self) -> CH_TX_ERR_R

Bit 6 - The interrupt raw bit for CHANNEL2. Triggered when error occurs.

pub fn ch3_tx_err(&self) -> CH_TX_ERR_R

Bit 7 - The interrupt raw bit for CHANNEL3. Triggered when error occurs.

pub fn ch_tx_thr_event(&self, n: u8) -> CH_TX_THR_EVENT_R

The interrupt raw bit for CHANNEL(0-3). Triggered when transmitter sent more data than configured value.

NOTE: n is number of field in register. n == 0 corresponds to CH0_TX_THR_EVENT field

pub fn ch_tx_thr_event_iter( &self ) -> impl Iterator<Item = CH_TX_THR_EVENT_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL(0-3). Triggered when transmitter sent more data than configured value.

pub fn ch0_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 8 - The interrupt raw bit for CHANNEL0. Triggered when transmitter sent more data than configured value.

pub fn ch1_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 9 - The interrupt raw bit for CHANNEL1. Triggered when transmitter sent more data than configured value.

pub fn ch2_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 10 - The interrupt raw bit for CHANNEL2. Triggered when transmitter sent more data than configured value.

pub fn ch3_tx_thr_event(&self) -> CH_TX_THR_EVENT_R

Bit 11 - The interrupt raw bit for CHANNEL3. Triggered when transmitter sent more data than configured value.

pub fn ch_tx_loop(&self, n: u8) -> CH_TX_LOOP_R

The interrupt raw bit for CHANNEL(0-3). Triggered when the loop count reaches the configured threshold value.

NOTE: n is number of field in register. n == 0 corresponds to CH0_TX_LOOP field

pub fn ch_tx_loop_iter(&self) -> impl Iterator<Item = CH_TX_LOOP_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL(0-3). Triggered when the loop count reaches the configured threshold value.

pub fn ch0_tx_loop(&self) -> CH_TX_LOOP_R

Bit 12 - The interrupt raw bit for CHANNEL0. Triggered when the loop count reaches the configured threshold value.

pub fn ch1_tx_loop(&self) -> CH_TX_LOOP_R

Bit 13 - The interrupt raw bit for CHANNEL1. Triggered when the loop count reaches the configured threshold value.

pub fn ch2_tx_loop(&self) -> CH_TX_LOOP_R

Bit 14 - The interrupt raw bit for CHANNEL2. Triggered when the loop count reaches the configured threshold value.

pub fn ch3_tx_loop(&self) -> CH_TX_LOOP_R

Bit 15 - The interrupt raw bit for CHANNEL3. Triggered when the loop count reaches the configured threshold value.

pub fn ch_rx_end(&self, n: u8) -> CH_RX_END_R

The interrupt raw bit for CHANNEL4. Triggered when reception done.

NOTE: n is number of field in register. n == 0 corresponds to CH4_RX_END field

pub fn ch_rx_end_iter(&self) -> impl Iterator<Item = CH_RX_END_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL4. Triggered when reception done.

pub fn ch4_rx_end(&self) -> CH_RX_END_R

Bit 16 - The interrupt raw bit for CHANNEL4. Triggered when reception done.

pub fn ch5_rx_end(&self) -> CH_RX_END_R

Bit 17 - The interrupt raw bit for CHANNEL4. Triggered when reception done.

pub fn ch6_rx_end(&self) -> CH_RX_END_R

Bit 18 - The interrupt raw bit for CHANNEL4. Triggered when reception done.

pub fn ch7_rx_end(&self) -> CH_RX_END_R

Bit 19 - The interrupt raw bit for CHANNEL4. Triggered when reception done.

pub fn ch_rx_err(&self, n: u8) -> CH_RX_ERR_R

The interrupt raw bit for CHANNEL4. Triggered when error occurs.

NOTE: n is number of field in register. n == 0 corresponds to CH4_RX_ERR field

pub fn ch_rx_err_iter(&self) -> impl Iterator<Item = CH_RX_ERR_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL4. Triggered when error occurs.

pub fn ch4_rx_err(&self) -> CH_RX_ERR_R

Bit 20 - The interrupt raw bit for CHANNEL4. Triggered when error occurs.

pub fn ch5_rx_err(&self) -> CH_RX_ERR_R

Bit 21 - The interrupt raw bit for CHANNEL4. Triggered when error occurs.

pub fn ch6_rx_err(&self) -> CH_RX_ERR_R

Bit 22 - The interrupt raw bit for CHANNEL4. Triggered when error occurs.

pub fn ch7_rx_err(&self) -> CH_RX_ERR_R

Bit 23 - The interrupt raw bit for CHANNEL4. Triggered when error occurs.

pub fn ch_rx_thr_event(&self, n: u8) -> CH_RX_THR_EVENT_R

The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

NOTE: n is number of field in register. n == 0 corresponds to CH4_RX_THR_EVENT field

pub fn ch_rx_thr_event_iter( &self ) -> impl Iterator<Item = CH_RX_THR_EVENT_R> + '_

Iterator for array of: The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

pub fn ch4_rx_thr_event(&self) -> CH_RX_THR_EVENT_R

Bit 24 - The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

pub fn ch5_rx_thr_event(&self) -> CH_RX_THR_EVENT_R

Bit 25 - The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

pub fn ch6_rx_thr_event(&self) -> CH_RX_THR_EVENT_R

Bit 26 - The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

pub fn ch7_rx_thr_event(&self) -> CH_RX_THR_EVENT_R

Bit 27 - The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

pub fn tx_ch3_dma_access_fail(&self) -> TX_CH3_DMA_ACCESS_FAIL_R

Bit 28 - The interrupt raw bit for CHANNEL3. Triggered when dma accessing CHANNEL3 fails.

pub fn rx_ch7_dma_access_fail(&self) -> RX_CH7_DMA_ACCESS_FAIL_R

Bit 29 - The interrupt raw bit for CHANNEL7. Triggered when dma accessing CHANNEL7 fails.