Module esp32s3::rtc_cntl::low_power_st

reserved register

Structs

LOW_POWER_ST_SPEC

reserved register

Type Aliases

COCPU_STATE_DONE_R

Field COCPU_STATE_DONE reader - ulp/cocpu is done

COCPU_STATE_SLP_R

Field COCPU_STATE_SLP reader - ulp/cocpu is in sleep state

COCPU_STATE_START_R

Field COCPU_STATE_START reader - ulp/cocpu should start to work

COCPU_STATE_SWITCH_R

Field COCPU_STATE_SWITCH reader - ulp/cocpu is about to working. Switch rtc main state

DIG_ISO_R

Field DIG_ISO reader - digital wrap iso

IN_LOW_POWER_STATE_R

Field IN_LOW_POWER_STATE reader - rtc main state machine is in the states of low power

IN_WAKEUP_STATE_R

Field IN_WAKEUP_STATE reader - rtc main state machine is in the states of wakeup process

MAIN_STATE_IN_IDLE_R

Field MAIN_STATE_IN_IDLE reader - rtc main state machine is in idle state

MAIN_STATE_IN_SLP_R

Field MAIN_STATE_IN_SLP reader - rtc main state machine is in sleep state

MAIN_STATE_IN_WAIT_8M_R

Field MAIN_STATE_IN_WAIT_8M reader - rtc main state machine is in wait 8m state

MAIN_STATE_IN_WAIT_PLL_R

Field MAIN_STATE_IN_WAIT_PLL reader - rtc main state machine is in wait pll state

MAIN_STATE_IN_WAIT_XTL_R

Field MAIN_STATE_IN_WAIT_XTL reader - rtc main state machine is in wait xtal state

MAIN_STATE_PLL_ON_R

Field MAIN_STATE_PLL_ON reader - rtc main state machine is in states that pll should be running

MAIN_STATE_R

Field MAIN_STATE reader - rtc main state machine status

MAIN_STATE_WAIT_END_R

Field MAIN_STATE_WAIT_END reader - rtc main state machine has been waited for some cycles

MAIN_STATE_XTAL_ISO_R

Field MAIN_STATE_XTAL_ISO reader - no use any more

PERI_ISO_R

Field PERI_ISO reader - rtc peripheral iso

R

Register LOW_POWER_ST reader

RDY_FOR_WAKEUP_R

Field RDY_FOR_WAKEUP reader - rtc is ready to receive wake up trigger from wake up source

TOUCH_STATE_DONE_R

Field TOUCH_STATE_DONE reader - touch is done

TOUCH_STATE_SLP_R

Field TOUCH_STATE_SLP reader - touch is in sleep state

TOUCH_STATE_START_R

Field TOUCH_STATE_START reader - touch should start to work

TOUCH_STATE_SWITCH_R

Field TOUCH_STATE_SWITCH reader - touch is about to working. Switch rtc main state

WIFI_ISO_R

Field WIFI_ISO reader - wifi iso

XPD_DIG_DCDC_R

Field XPD_DIG_DCDC reader - External DCDC power down

XPD_DIG_R

Field XPD_DIG reader - digital wrap power down

XPD_ROM0_R

Field XPD_ROM0 reader - rom0 power down

XPD_RTC_PERI_R

Field XPD_RTC_PERI reader - rtc peripheral power down

XPD_WIFI_R

Field XPD_WIFI reader - wifi wrap power down