Type Alias esp32s3::rtc_cntl::rtc::R

pub type R = R<RTC_SPEC>;

Register RTC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn dig_reg_cal_en(&self) -> DIG_REG_CAL_EN_R

Bit 7 - enable dig regulator cali

pub fn sck_dcap(&self) -> SCK_DCAP_R

Bits 14:21 - SCK_DCAP

pub fn dboost_force_pd(&self) -> DBOOST_FORCE_PD_R

Bit 28 - RTC_DBOOST force power down

pub fn dboost_force_pu(&self) -> DBOOST_FORCE_PU_R

Bit 29 - RTC_DBOOST force power up

pub fn regulator_force_pd(&self) -> REGULATOR_FORCE_PD_R

Bit 30 - RTC_REG force power down (for RTC_REG power down means decrease the voltage to 0.8v or lower )

pub fn regulator_force_pu(&self) -> REGULATOR_FORCE_PU_R

Bit 31 - RTC_REG force power on (for RTC_REG power down means decrease the voltage to 0.8v or lower )