Module esp32s3::spi0::ctrl2

SPI0 control 2 register.

Structs

CTRL2_SPEC

SPI0 control 2 register.

Type Aliases

CS_HOLD_DELAY_R

Field CS_HOLD_DELAY reader - These bits are used to set the minimum CS high time tSHSL between SPI burst transfer when accesses to flash. tSHSL is (SPI_MEM_CS_HOLD_DELAY[5:0] + 1) MSPI core clock cycles.

CS_HOLD_DELAY_W

Field CS_HOLD_DELAY writer - These bits are used to set the minimum CS high time tSHSL between SPI burst transfer when accesses to flash. tSHSL is (SPI_MEM_CS_HOLD_DELAY[5:0] + 1) MSPI core clock cycles.

CS_HOLD_TIME_R

Field CS_HOLD_TIME reader - SPI Bus CS (SPI_CS) signal is delayed to inactive by SPI Bus clock (SPI_CLK), which is the SPI_CS hold time in non-ECC mode. These bits are combined with SPI_MEM_CS_HOLD bit.

CS_HOLD_TIME_W

Field CS_HOLD_TIME writer - SPI Bus CS (SPI_CS) signal is delayed to inactive by SPI Bus clock (SPI_CLK), which is the SPI_CS hold time in non-ECC mode. These bits are combined with SPI_MEM_CS_HOLD bit.

CS_SETUP_TIME_R

Field CS_SETUP_TIME reader - (cycles-1) of PREP phase by SPI_CLK, which is the SPI_CS setup time. These bits are combined with SPI_MEM_CS_SETUP bit.

CS_SETUP_TIME_W

Field CS_SETUP_TIME writer - (cycles-1) of PREP phase by SPI_CLK, which is the SPI_CS setup time. These bits are combined with SPI_MEM_CS_SETUP bit.

ECC_16TO18_BYTE_EN_R

Field ECC_16TO18_BYTE_EN reader - Set this bit to enable MSPI ECC 16 bytes data with 2 ECC bytes mode when accesses flash.

ECC_16TO18_BYTE_EN_W

Field ECC_16TO18_BYTE_EN writer - Set this bit to enable MSPI ECC 16 bytes data with 2 ECC bytes mode when accesses flash.

ECC_CS_HOLD_TIME_R

Field ECC_CS_HOLD_TIME reader - SPI_MEM_CS_HOLD_TIME + SPI_MEM_ECC_CS_HOLD_TIME is the SPI_CS hold cycle in ECC mode when accessed flash.

ECC_CS_HOLD_TIME_W

Field ECC_CS_HOLD_TIME writer - SPI_MEM_CS_HOLD_TIME + SPI_MEM_ECC_CS_HOLD_TIME is the SPI_CS hold cycle in ECC mode when accessed flash.

ECC_SKIP_PAGE_CORNER_R

Field ECC_SKIP_PAGE_CORNER reader - 1: MSPI skips page corner when accesses flash. 0: Not skip page corner when accesses flash.

ECC_SKIP_PAGE_CORNER_W

Field ECC_SKIP_PAGE_CORNER writer - 1: MSPI skips page corner when accesses flash. 0: Not skip page corner when accesses flash.

R

Register CTRL2 reader

SYNC_RESET_R

Field SYNC_RESET reader - The FSM will be reset.

SYNC_RESET_W

Field SYNC_RESET writer - The FSM will be reset.

W

Register CTRL2 writer