Type Alias esp32s3::spi0::spi_smem_ddr::W

pub type W = W<SPI_SMEM_DDR_SPEC>;

Register SPI_SMEM_DDR writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn en(&mut self) -> EN_W<'_, SPI_SMEM_DDR_SPEC>

Bit 0 - 1: in ddr mode, 0 in sdr mode

pub fn spi_smem_var_dummy( &mut self ) -> SPI_SMEM_VAR_DUMMY_W<'_, SPI_SMEM_DDR_SPEC>

Bit 1 - Set the bit to enable variable dummy cycle in spi ddr mode.

pub fn rdat_swp(&mut self) -> RDAT_SWP_W<'_, SPI_SMEM_DDR_SPEC>

Bit 2 - Set the bit to reorder rx data of the word in spi ddr mode.

pub fn wdat_swp(&mut self) -> WDAT_SWP_W<'_, SPI_SMEM_DDR_SPEC>

Bit 3 - Set the bit to reorder tx data of the word in spi ddr mode.

pub fn cmd_dis(&mut self) -> CMD_DIS_W<'_, SPI_SMEM_DDR_SPEC>

Bit 4 - the bit is used to disable dual edge in CMD phase when ddr mode.

pub fn spi_smem_outminbytelen( &mut self ) -> SPI_SMEM_OUTMINBYTELEN_W<'_, SPI_SMEM_DDR_SPEC>

Bits 5:11 - It is the minimum output data length in the ddr psram.

pub fn spi_smem_tx_ddr_msk_en( &mut self ) -> SPI_SMEM_TX_DDR_MSK_EN_W<'_, SPI_SMEM_DDR_SPEC>

Bit 12 - Set this bit to mask the first or the last byte in MSPI ECC DDR write mode, when accesses to external RAM.

pub fn spi_smem_rx_ddr_msk_en( &mut self ) -> SPI_SMEM_RX_DDR_MSK_EN_W<'_, SPI_SMEM_DDR_SPEC>

Bit 13 - Set this bit to mask the first or the last byte in MSPI ECC DDR read mode, when accesses to external RAM.

pub fn spi_smem_usr_ddr_dqs_thd( &mut self ) -> SPI_SMEM_USR_DDR_DQS_THD_W<'_, SPI_SMEM_DDR_SPEC>

Bits 14:20 - The delay number of data strobe which from memory based on SPI_CLK.

pub fn dqs_loop(&mut self) -> DQS_LOOP_W<'_, SPI_SMEM_DDR_SPEC>

Bit 21 - 1: Use internal signal as data strobe, the strobe can not be delayed by input timing module. 0: Use input SPI_DQS signal from PAD as data strobe, the strobe can be delayed by input timing module

pub fn dqs_loop_mode(&mut self) -> DQS_LOOP_MODE_W<'_, SPI_SMEM_DDR_SPEC>

Bit 22 - When SPI_SMEM_DDR_DQS_LOOP and SPI_SMEM_DDR_EN are set, 1: Use internal SPI_CLK as data strobe. 0: Use internal ~SPI_CLK as data strobe. Otherwise this bit is not active.

pub fn spi_smem_clk_diff_en( &mut self ) -> SPI_SMEM_CLK_DIFF_EN_W<'_, SPI_SMEM_DDR_SPEC>

Bit 24 - Set this bit to enable the differential SPI_CLK#.

pub fn spi_smem_hyperbus_mode( &mut self ) -> SPI_SMEM_HYPERBUS_MODE_W<'_, SPI_SMEM_DDR_SPEC>

Bit 25 - Set this bit to enable the SPI HyperBus mode.

pub fn spi_smem_dqs_ca_in( &mut self ) -> SPI_SMEM_DQS_CA_IN_W<'_, SPI_SMEM_DDR_SPEC>

Bit 26 - Set this bit to enable the input of SPI_DQS signal in SPI phases of CMD and ADDR.

pub fn spi_smem_hyperbus_dummy_2x( &mut self ) -> SPI_SMEM_HYPERBUS_DUMMY_2X_W<'_, SPI_SMEM_DDR_SPEC>

Bit 27 - Set this bit to enable the vary dummy function in SPI HyperBus mode, when SPI0 accesses to flash or SPI1 accesses flash or sram.

pub fn spi_smem_clk_diff_inv( &mut self ) -> SPI_SMEM_CLK_DIFF_INV_W<'_, SPI_SMEM_DDR_SPEC>

Bit 28 - Set this bit to invert SPI_DIFF when accesses to external RAM. .

pub fn spi_smem_octa_ram_addr( &mut self ) -> SPI_SMEM_OCTA_RAM_ADDR_W<'_, SPI_SMEM_DDR_SPEC>

Bit 29 - Set this bit to enable octa_ram address out when accesses to external RAM, which means ADDR_OUT[31:0] = {spi_usr_addr_value[25:4], 6’d0, spi_usr_addr_value[3:1], 1’b0}.

pub fn spi_smem_hyperbus_ca( &mut self ) -> SPI_SMEM_HYPERBUS_CA_W<'_, SPI_SMEM_DDR_SPEC>

Bit 30 - Set this bit to enable HyperRAM address out when accesses to external RAM, which means ADDR_OUT[31:0] = {spi_usr_addr_value[19:4], 13’d0, spi_usr_addr_value[3:1]}.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual