Type Alias esp32s3::spi1::flash_sus_cmd::W

pub type W = W<FLASH_SUS_CMD_SPEC>;

Register FLASH_SUS_CMD writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn flash_per(&mut self) -> FLASH_PER_W<'_, FLASH_SUS_CMD_SPEC>

Bit 0 - program erase resume bit, program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn flash_pes(&mut self) -> FLASH_PES_W<'_, FLASH_SUS_CMD_SPEC>

Bit 1 - program erase suspend bit, program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn flash_per_wait_en( &mut self ) -> FLASH_PER_WAIT_EN_W<'_, FLASH_SUS_CMD_SPEC>

Bit 2 - Set this bit to add delay time after program erase resume(PER) is sent.

pub fn flash_pes_wait_en( &mut self ) -> FLASH_PES_WAIT_EN_W<'_, FLASH_SUS_CMD_SPEC>

Bit 3 - Set this bit to add delay time after program erase suspend(PES) command is sent.

pub fn pes_per_en(&mut self) -> PES_PER_EN_W<'_, FLASH_SUS_CMD_SPEC>

Bit 4 - Set this bit to enable PES transfer trigger PES transfer option.

pub fn pesr_idle_en(&mut self) -> PESR_IDLE_EN_W<'_, FLASH_SUS_CMD_SPEC>

Bit 5 - 1: Separate PER flash wait idle and PES flash wait idle. 0: Not separate.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual