Module esp32s3::spi1::user

SPI1 user register.

Structs

USER_SPEC

SPI1 user register.

Type Aliases

CK_OUT_EDGE_R

Field CK_OUT_EDGE reader - This bit, combined with SPI_MEM_CK_IDLE_EDGE bit, is used to change the clock mode 0~3 of SPI_CLK.

CK_OUT_EDGE_W

Field CK_OUT_EDGE writer - This bit, combined with SPI_MEM_CK_IDLE_EDGE bit, is used to change the clock mode 0~3 of SPI_CLK.

FWRITE_DIO_R

Field FWRITE_DIO reader - Set this bit to enable 2-bm in ADDR and DOUT phase in SPI1 write operation.

FWRITE_DIO_W

Field FWRITE_DIO writer - Set this bit to enable 2-bm in ADDR and DOUT phase in SPI1 write operation.

FWRITE_DUAL_R

Field FWRITE_DUAL reader - Set this bit to enable 2-bm in DOUT phase in SPI1 write operation.

FWRITE_DUAL_W

Field FWRITE_DUAL writer - Set this bit to enable 2-bm in DOUT phase in SPI1 write operation.

FWRITE_QIO_R

Field FWRITE_QIO reader - Set this bit to enable 4-bit-mode(4-bm) in ADDR and DOUT phase in SPI1 write operation.

FWRITE_QIO_W

Field FWRITE_QIO writer - Set this bit to enable 4-bit-mode(4-bm) in ADDR and DOUT phase in SPI1 write operation.

FWRITE_QUAD_R

Field FWRITE_QUAD reader - Set this bit to enable 4-bm in DOUT phase in SPI1 write operation.

FWRITE_QUAD_W

Field FWRITE_QUAD writer - Set this bit to enable 4-bm in DOUT phase in SPI1 write operation.

R

Register USER reader

USR_ADDR_R

Field USR_ADDR reader - Set this bit to enable enable the ADDR phase of an operation.

USR_ADDR_W

Field USR_ADDR writer - Set this bit to enable enable the ADDR phase of an operation.

USR_COMMAND_R

Field USR_COMMAND reader - Set this bit to enable enable the CMD phase of an operation.

USR_COMMAND_W

Field USR_COMMAND writer - Set this bit to enable enable the CMD phase of an operation.

USR_DUMMY_IDLE_R

Field USR_DUMMY_IDLE reader - SPI_CLK is disabled(No clock edges) in DUMMY phase when the bit is enable.

USR_DUMMY_IDLE_W

Field USR_DUMMY_IDLE writer - SPI_CLK is disabled(No clock edges) in DUMMY phase when the bit is enable.

USR_DUMMY_R

Field USR_DUMMY reader - Set this bit to enable enable the DUMMY phase of an operation.

USR_DUMMY_W

Field USR_DUMMY writer - Set this bit to enable enable the DUMMY phase of an operation.

USR_MISO_HIGHPART_R

Field USR_MISO_HIGHPART reader - DIN phase only access to high-part of the buffer SPI_MEM_W8_REG~SPI_MEM_W15_REG. 1: enable 0: disable.

USR_MISO_HIGHPART_W

Field USR_MISO_HIGHPART writer - DIN phase only access to high-part of the buffer SPI_MEM_W8_REG~SPI_MEM_W15_REG. 1: enable 0: disable.

USR_MISO_R

Field USR_MISO reader - Set this bit to enable enable the DIN phase of a read-data operation.

USR_MISO_W

Field USR_MISO writer - Set this bit to enable enable the DIN phase of a read-data operation.

USR_MOSI_HIGHPART_R

Field USR_MOSI_HIGHPART reader - DOUT phase only access to high-part of the buffer SPI_MEM_W8_REG~SPI_MEM_W15_REG. 1: enable 0: disable.

USR_MOSI_HIGHPART_W

Field USR_MOSI_HIGHPART writer - DOUT phase only access to high-part of the buffer SPI_MEM_W8_REG~SPI_MEM_W15_REG. 1: enable 0: disable.

USR_MOSI_R

Field USR_MOSI reader - Set this bit to enable the DOUT phase of an write-data operation.

USR_MOSI_W

Field USR_MOSI writer - Set this bit to enable the DOUT phase of an write-data operation.

W

Register USER writer