Type Alias esp32s3::twai0::status::R

pub type R = R<STATUS_SPEC>;

Register STATUS reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rx_buf_st(&self) -> RX_BUF_ST_R

Bit 0 - 1: The data in the RX buffer is not empty, with at least one received data packet.

pub fn overrun_st(&self) -> OVERRUN_ST_R

Bit 1 - 1: The RX FIFO is full and data overrun has occurred.

pub fn tx_buf_st(&self) -> TX_BUF_ST_R

Bit 2 - 1: The TX buffer is empty, the CPU may write a message into it.

pub fn tx_complete(&self) -> TX_COMPLETE_R

Bit 3 - 1: The TWAI controller has successfully received a packet from the bus.

pub fn rx_st(&self) -> RX_ST_R

Bit 4 - 1: The TWAI Controller is receiving a message from the bus.

pub fn tx_st(&self) -> TX_ST_R

Bit 5 - 1: The TWAI Controller is transmitting a message to the bus.

pub fn err_st(&self) -> ERR_ST_R

Bit 6 - 1: At least one of the RX/TX error counter has reached or exceeded the value set in register TWAI_ERR_WARNING_LIMIT_REG.

pub fn bus_off_st(&self) -> BUS_OFF_ST_R

Bit 7 - 1: In bus-off status, the TWAI Controller is no longer involved in bus activities.

pub fn miss_st(&self) -> MISS_ST_R

Bit 8 - This bit reflects whether the data packet in the RX FIFO is complete. 1: The current packet is missing; 0: The current packet is complete