Module esp32s3::usb_wrap::otg_conf

USB OTG Wrapper Configure Register

Structs

OTG_CONF_SPEC

USB OTG Wrapper Configure Register

Type Aliases

AHB_CLK_FORCE_ON_R

Field AHB_CLK_FORCE_ON reader - Force ahb clock always on

AHB_CLK_FORCE_ON_W

Field AHB_CLK_FORCE_ON writer - Force ahb clock always on

CLK_EN_R

Field CLK_EN reader - Disable auto clock gating of CSR registers

CLK_EN_W

Field CLK_EN writer - Disable auto clock gating of CSR registers

DBNCE_FLTR_BYPASS_R

Field DBNCE_FLTR_BYPASS reader - Bypass Debounce filters for avalid,bvalid,vbusvalid,session end, id signals

DBNCE_FLTR_BYPASS_W

Field DBNCE_FLTR_BYPASS writer - Bypass Debounce filters for avalid,bvalid,vbusvalid,session end, id signals

DFIFO_FORCE_PD_R

Field DFIFO_FORCE_PD reader - Force the dfifo to go into low power mode. The data in dfifo will not lost.

DFIFO_FORCE_PD_W

Field DFIFO_FORCE_PD writer - Force the dfifo to go into low power mode. The data in dfifo will not lost.

DFIFO_FORCE_PU_R

Field DFIFO_FORCE_PU reader - Disable the dfifo to go into low power mode. The data in dfifo will not lost.

DFIFO_FORCE_PU_W

Field DFIFO_FORCE_PU writer - Disable the dfifo to go into low power mode. The data in dfifo will not lost.

DM_PULLDOWN_R

Field DM_PULLDOWN reader - Controlle USB D+ pulldown

DM_PULLDOWN_W

Field DM_PULLDOWN writer - Controlle USB D+ pulldown

DM_PULLUP_R

Field DM_PULLUP reader - Controlle USB D+ pullup

DM_PULLUP_W

Field DM_PULLUP writer - Controlle USB D+ pullup

DP_PULLDOWN_R

Field DP_PULLDOWN reader - Controlle USB D+ pulldown

DP_PULLDOWN_W

Field DP_PULLDOWN writer - Controlle USB D+ pulldown

DP_PULLUP_R

Field DP_PULLUP reader - Controlle USB D+ pullup

DP_PULLUP_W

Field DP_PULLUP writer - Controlle USB D+ pullup

EXCHG_PINS_OVERRIDE_R

Field EXCHG_PINS_OVERRIDE reader - Enable software controlle USB D+ D- exchange

EXCHG_PINS_OVERRIDE_W

Field EXCHG_PINS_OVERRIDE writer - Enable software controlle USB D+ D- exchange

EXCHG_PINS_R

Field EXCHG_PINS reader - USB D+ D- exchange. 1’b0: don’t change. 1’b1: exchange D+ D-

EXCHG_PINS_W

Field EXCHG_PINS writer - USB D+ D- exchange. 1’b0: don’t change. 1’b1: exchange D+ D-

PAD_PULL_OVERRIDE_R

Field PAD_PULL_OVERRIDE reader - Enable software controlle USB D+ D- pullup pulldown

PAD_PULL_OVERRIDE_W

Field PAD_PULL_OVERRIDE writer - Enable software controlle USB D+ D- pullup pulldown

PHY_CLK_FORCE_ON_R

Field PHY_CLK_FORCE_ON reader - Force phy clock always on

PHY_CLK_FORCE_ON_W

Field PHY_CLK_FORCE_ON writer - Force phy clock always on

PHY_SEL_R

Field PHY_SEL reader - Select internal external PHY. 1’b0: Select internal PHY. 1’b1: Select external PHY.

PHY_SEL_W

Field PHY_SEL writer - Select internal external PHY. 1’b0: Select internal PHY. 1’b1: Select external PHY.

PHY_TX_EDGE_SEL_R

Field PHY_TX_EDGE_SEL reader - Select phy tx signal output clock edge. 1’b0: negedge. 1’b1: posedge.

PHY_TX_EDGE_SEL_W

Field PHY_TX_EDGE_SEL writer - Select phy tx signal output clock edge. 1’b0: negedge. 1’b1: posedge.

PULLUP_VALUE_R

Field PULLUP_VALUE reader - Controlle pullup value. 1’b0: typical value is 2.4K. 1’b1: typical value is 1.2K.

PULLUP_VALUE_W

Field PULLUP_VALUE writer - Controlle pullup value. 1’b0: typical value is 2.4K. 1’b1: typical value is 1.2K.

R

Register OTG_CONF reader

SRP_SESSEND_OVERRIDE_R

Field SRP_SESSEND_OVERRIDE reader - This bit is used to enable the software over-ride of srp session end signal. 1’b0: the signal is controlled by the chip input. 1’b1: the signal is controlled by the software.

SRP_SESSEND_OVERRIDE_W

Field SRP_SESSEND_OVERRIDE writer - This bit is used to enable the software over-ride of srp session end signal. 1’b0: the signal is controlled by the chip input. 1’b1: the signal is controlled by the software.

SRP_SESSEND_VALUE_R

Field SRP_SESSEND_VALUE reader - Software over-ride value of srp session end signal.

SRP_SESSEND_VALUE_W

Field SRP_SESSEND_VALUE writer - Software over-ride value of srp session end signal.

USB_PAD_ENABLE_R

Field USB_PAD_ENABLE reader - Enable USB pad function

USB_PAD_ENABLE_W

Field USB_PAD_ENABLE writer - Enable USB pad function

VREFH_R

Field VREFH reader - Control single-end input high threshold,1.76V to 2V, step 80mV

VREFH_W

Field VREFH writer - Control single-end input high threshold,1.76V to 2V, step 80mV

VREFL_R

Field VREFL reader - Control single-end input low threshold,0.8V to 1.04V, step 80mV

VREFL_W

Field VREFL writer - Control single-end input low threshold,0.8V to 1.04V, step 80mV

VREF_OVERRIDE_R

Field VREF_OVERRIDE reader - Enable software controlle input threshold

VREF_OVERRIDE_W

Field VREF_OVERRIDE writer - Enable software controlle input threshold

W

Register OTG_CONF writer