Type Alias esp32s3::usb_wrap::otg_conf::W

pub type W = W<OTG_CONF_SPEC>;

Register OTG_CONF writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn srp_sessend_override( &mut self ) -> SRP_SESSEND_OVERRIDE_W<'_, OTG_CONF_SPEC>

Bit 0 - This bit is used to enable the software over-ride of srp session end signal. 1’b0: the signal is controlled by the chip input. 1’b1: the signal is controlled by the software.

pub fn srp_sessend_value(&mut self) -> SRP_SESSEND_VALUE_W<'_, OTG_CONF_SPEC>

Bit 1 - Software over-ride value of srp session end signal.

pub fn phy_sel(&mut self) -> PHY_SEL_W<'_, OTG_CONF_SPEC>

Bit 2 - Select internal external PHY. 1’b0: Select internal PHY. 1’b1: Select external PHY.

pub fn dfifo_force_pd(&mut self) -> DFIFO_FORCE_PD_W<'_, OTG_CONF_SPEC>

Bit 3 - Force the dfifo to go into low power mode. The data in dfifo will not lost.

pub fn dbnce_fltr_bypass(&mut self) -> DBNCE_FLTR_BYPASS_W<'_, OTG_CONF_SPEC>

Bit 4 - Bypass Debounce filters for avalid,bvalid,vbusvalid,session end, id signals

pub fn exchg_pins_override( &mut self ) -> EXCHG_PINS_OVERRIDE_W<'_, OTG_CONF_SPEC>

Bit 5 - Enable software controlle USB D+ D- exchange

pub fn exchg_pins(&mut self) -> EXCHG_PINS_W<'_, OTG_CONF_SPEC>

Bit 6 - USB D+ D- exchange. 1’b0: don’t change. 1’b1: exchange D+ D-

pub fn vrefh(&mut self) -> VREFH_W<'_, OTG_CONF_SPEC>

Bits 7:8 - Control single-end input high threshold,1.76V to 2V, step 80mV

pub fn vrefl(&mut self) -> VREFL_W<'_, OTG_CONF_SPEC>

Bits 9:10 - Control single-end input low threshold,0.8V to 1.04V, step 80mV

pub fn vref_override(&mut self) -> VREF_OVERRIDE_W<'_, OTG_CONF_SPEC>

Bit 11 - Enable software controlle input threshold

pub fn pad_pull_override(&mut self) -> PAD_PULL_OVERRIDE_W<'_, OTG_CONF_SPEC>

Bit 12 - Enable software controlle USB D+ D- pullup pulldown

pub fn dp_pullup(&mut self) -> DP_PULLUP_W<'_, OTG_CONF_SPEC>

Bit 13 - Controlle USB D+ pullup

pub fn dp_pulldown(&mut self) -> DP_PULLDOWN_W<'_, OTG_CONF_SPEC>

Bit 14 - Controlle USB D+ pulldown

pub fn dm_pullup(&mut self) -> DM_PULLUP_W<'_, OTG_CONF_SPEC>

Bit 15 - Controlle USB D+ pullup

pub fn dm_pulldown(&mut self) -> DM_PULLDOWN_W<'_, OTG_CONF_SPEC>

Bit 16 - Controlle USB D+ pulldown

pub fn pullup_value(&mut self) -> PULLUP_VALUE_W<'_, OTG_CONF_SPEC>

Bit 17 - Controlle pullup value. 1’b0: typical value is 2.4K. 1’b1: typical value is 1.2K.

pub fn usb_pad_enable(&mut self) -> USB_PAD_ENABLE_W<'_, OTG_CONF_SPEC>

Bit 18 - Enable USB pad function

pub fn ahb_clk_force_on(&mut self) -> AHB_CLK_FORCE_ON_W<'_, OTG_CONF_SPEC>

Bit 19 - Force ahb clock always on

pub fn phy_clk_force_on(&mut self) -> PHY_CLK_FORCE_ON_W<'_, OTG_CONF_SPEC>

Bit 20 - Force phy clock always on

pub fn phy_tx_edge_sel(&mut self) -> PHY_TX_EDGE_SEL_W<'_, OTG_CONF_SPEC>

Bit 21 - Select phy tx signal output clock edge. 1’b0: negedge. 1’b1: posedge.

pub fn dfifo_force_pu(&mut self) -> DFIFO_FORCE_PU_W<'_, OTG_CONF_SPEC>

Bit 22 - Disable the dfifo to go into low power mode. The data in dfifo will not lost.

pub fn clk_en(&mut self) -> CLK_EN_W<'_, OTG_CONF_SPEC>

Bit 31 - Disable auto clock gating of CSR registers

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual